+++ pkg/storage/stores/series/index/table_manager.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/grafana/dskit/instrument"
 	"github.com/grafana/dskit/mtime"
@@ -21,7 +22,6 @@ import (
 
 	"github.com/grafana/loki/pkg/storage/config"
 	"github.com/grafana/loki/pkg/util/constants"
-	util_log "github.com/grafana/loki/pkg/util/log"
 )
 
 const (
@@ -161,6 +161,7 @@ type TableManager struct {
 
 	client       TableClient
 	cfg          TableManagerConfig
+	logger       log.Logger
 	schemaCfg    config.SchemaConfig
 	maxChunkAge  time.Duration
 	bucketClient BucketClient
@@ -172,7 +173,7 @@ type TableManager struct {
 
 // NewTableManager makes a new TableManager
 func NewTableManager(cfg TableManagerConfig, schemaCfg config.SchemaConfig, maxChunkAge time.Duration, tableClient TableClient,
-	objectClient BucketClient, extraTables []ExtraTables, registerer prometheus.Registerer,
+	objectClient BucketClient, extraTables []ExtraTables, registerer prometheus.Registerer, logger log.Logger,
 ) (*TableManager, error) {
 	if cfg.RetentionPeriod != 0 {
 		// Assume the newest config is the one to use for validation of retention
@@ -184,6 +185,7 @@ func NewTableManager(cfg TableManagerConfig, schemaCfg config.SchemaConfig, maxC
 
 	tm := &TableManager{
 		cfg:          cfg,
+		logger:       logger,
 		schemaCfg:    schemaCfg,
 		maxChunkAge:  maxChunkAge,
 		client:       tableClient,
@@ -221,7 +223,7 @@ func (m *TableManager) loop(ctx context.Context) error {
 	if err := instrument.CollectedRequest(context.Background(), "TableManager.SyncTables", instrument.NewHistogramCollector(m.metrics.syncTableDuration), instrument.ErrorCode, func(ctx context.Context) error {
 		return m.SyncTables(ctx)
 	}); err != nil {
-		level.Error(util_log.Logger).Log("msg", "error syncing tables", "err", err)
+		level.Error(m.logger).Log("msg", "error syncing tables", "err", err)
 	}
 
 	// Sleep for a bit to spread the sync load across different times if the tablemanagers are all started at once.
@@ -237,7 +239,7 @@ func (m *TableManager) loop(ctx context.Context) error {
 			if err := instrument.CollectedRequest(context.Background(), "TableManager.SyncTables", instrument.NewHistogramCollector(m.metrics.syncTableDuration), instrument.ErrorCode, func(ctx context.Context) error {
 				return m.SyncTables(ctx)
 			}); err != nil {
-				level.Error(util_log.Logger).Log("msg", "error syncing tables", "err", err)
+				level.Error(m.logger).Log("msg", "error syncing tables", "err", err)
 			}
 		case <-ctx.Done():
 			return nil
@@ -260,7 +262,7 @@ func (m *TableManager) checkAndCreateExtraTables() error {
 		for _, tableDesc := range extraTables.Tables {
 			if _, ok := existingTablesMap[tableDesc.Name]; !ok {
 				// creating table
-				level.Info(util_log.Logger).Log("msg", "creating extra table",
+				level.Info(m.logger).Log("msg", "creating extra table",
 					"tableName", tableDesc.Name,
 					"provisionedRead", tableDesc.ProvisionedRead,
 					"provisionedWrite", tableDesc.ProvisionedWrite,
@@ -278,7 +280,7 @@ func (m *TableManager) checkAndCreateExtraTables() error {
 				continue
 			}
 
-			level.Info(util_log.Logger).Log("msg", "checking throughput of extra table", "table", tableDesc.Name)
+			level.Info(m.logger).Log("msg", "checking throughput of extra table", "table", tableDesc.Name)
 			// table already exists, lets check actual throughput for tables is same as what is in configurations, if not let us update it
 			current, _, err := extraTables.TableClient.DescribeTable(context.Background(), tableDesc.Name)
 			if err != nil {
@@ -286,7 +288,7 @@ func (m *TableManager) checkAndCreateExtraTables() error {
 			}
 
 			if !current.Equals(tableDesc) {
-				level.Info(util_log.Logger).Log("msg", "updating throughput of extra table",
+				level.Info(m.logger).Log("msg", "updating throughput of extra table",
 					"table", tableDesc.Name,
 					"tableName", tableDesc.Name,
 					"provisionedRead", tableDesc.ProvisionedRead,
@@ -310,7 +312,7 @@ func (m *TableManager) checkAndCreateExtraTables() error {
 func (m *TableManager) bucketRetentionIteration(ctx context.Context) error {
 	err := m.bucketClient.DeleteChunksBefore(ctx, mtime.Now().Add(-m.cfg.RetentionPeriod))
 	if err != nil {
-		level.Error(util_log.Logger).Log("msg", "error enforcing filesystem retention", "err", err)
+		level.Error(m.logger).Log("msg", "error enforcing filesystem retention", "err", err)
 	}
 
 	// don't return error, otherwise timer service would stop.
@@ -326,7 +328,7 @@ func (m *TableManager) SyncTables(ctx context.Context) error {
 	}
 
 	expected := m.calculateExpectedTables()
-	level.Debug(util_log.Logger).Log("msg", "syncing tables", "expected_tables", len(expected))
+	level.Debug(m.logger).Log("msg", "syncing tables", "expected_tables", len(expected))
 
 	toCreate, toCheckThroughput, toDelete, err := m.partitionTables(ctx, expected)
 	if err != nil {
@@ -477,7 +479,7 @@ func (m *TableManager) createTables(ctx context.Context, descriptions []config.T
 	merr := tsdb_errors.NewMulti()
 
 	for _, desc := range descriptions {
-		level.Debug(util_log.Logger).Log("msg", "creating table", "table", desc.Name)
+		level.Debug(m.logger).Log("msg", "creating table", "table", desc.Name)
 		err := m.client.CreateTable(ctx, desc)
 		if err != nil {
 			numFailures++
@@ -494,12 +496,12 @@ func (m *TableManager) deleteTables(ctx context.Context, descriptions []config.T
 	merr := tsdb_errors.NewMulti()
 
 	for _, desc := range descriptions {
-		level.Info(util_log.Logger).Log("msg", "table has exceeded the retention period", "table", desc.Name)
+		level.Info(m.logger).Log("msg", "table has exceeded the retention period", "table", desc.Name)
 		if !m.cfg.RetentionDeletesEnabled {
 			continue
 		}
 
-		level.Info(util_log.Logger).Log("msg", "deleting table", "table", desc.Name)
+		level.Info(m.logger).Log("msg", "deleting table", "table", desc.Name)
 		err := m.client.DeleteTable(ctx, desc.Name)
 		if err != nil {
 			numFailures++
@@ -513,7 +515,7 @@ func (m *TableManager) deleteTables(ctx context.Context, descriptions []config.T
 
 func (m *TableManager) updateTables(ctx context.Context, descriptions []config.TableDesc) error {
 	for _, expected := range descriptions {
-		level.Debug(util_log.Logger).Log("msg", "checking provisioned throughput on table", "table", expected.Name)
+		level.Debug(m.logger).Log("msg", "checking provisioned throughput on table", "table", expected.Name)
 		current, isActive, err := m.client.DescribeTable(ctx, expected.Name)
 		if err != nil {
 			return err
@@ -527,12 +529,12 @@ func (m *TableManager) updateTables(ctx context.Context, descriptions []config.T
 		}
 
 		if !isActive {
-			level.Info(util_log.Logger).Log("msg", "skipping update on table, not yet ACTIVE", "table", expected.Name)
+			level.Info(m.logger).Log("msg", "skipping update on table, not yet ACTIVE", "table", expected.Name)
 			continue
 		}
 
 		if expected.Equals(current) {
-			level.Info(util_log.Logger).Log("msg", "provisioned throughput on table, skipping", "table", current.Name, "read", current.ProvisionedRead, "write", current.ProvisionedWrite)
+			level.Info(m.logger).Log("msg", "provisioned throughput on table, skipping", "table", current.Name, "read", current.ProvisionedRead, "write", current.ProvisionedWrite)
 			continue
 		}